Trait linear_sumcheck::data_structures::ml_extension::MLExtension

pub trait MLExtension<F>: Sized where
    F: Field, {
    type BinaryArg: From<usize> + Into<usize> + Copy;
    type Error: Error + From<Error> + Display;
    fn num_variables(&self) -> Result<usize, Self::Error>;

    fn eval_binary(&self, point: Self::BinaryArg) -> Result<F, Self::Error>;

    fn eval_at(&self, point: &[F]) -> Result<F, Self::Error>;

    fn table(&self) -> Result<Vec<F>, Self::Error>;
}

Multilinear Extension

A data-structure representing a polynomial that is a multi-linear extension of an array.

Associated Types

type BinaryArg: From<usize> + Into<usize> + Copy

Point in Binary Form

This type represents a number viewed as a little endian binary string viewed as a point in {0,1}^L.

Example

0b1011 represents P(1,1,0,1)

type Error: Error + From<Error> + Display

Error Type

Loading content...
Required methods

fn num_variables(&self) -> Result<usize, Self::Error>

Number of variables (L)

This function returns the total number of variables in this polynomial. The number of variables is log2(size of array).

fn eval_binary(&self, point: Self::BinaryArg) -> Result<F, Self::Error>

Evaluate the polynomial given point in length-L binary string

This function takes a binary point and evaluate the polynomial at that point. This is equivalent to accessing an element in the array by index. Learn more about how the binary argument work in the doc of MLExtensionArray

fn eval_at(&self, point: &[F]) -> Result<F, Self::Error>

Evaluate the polynomial at a point in Field

fn table(&self) -> Result<Vec<F>, Self::Error>

Get the copy of the values of all evaluations. Index: binary argument, value: F

Loading content...