Trait linear_sumcheck::data_structures::ml_extension::SparseMLExtension

pub trait SparseMLExtension<F>: MLExtension<F> where
    F: Field, {
    fn sparse_table(&self) -> Result<Vec<(Self::BinaryArg, F)>, Self::Error>;
}

Sparse Multilinear Extension

Required methods

fn sparse_table(&self) -> Result<Vec<(Self::BinaryArg, F)>, Self::Error>

get the copy of the values of all evaluations that are not zero along with the index.

return: index as binary arg, value

Loading content...