Trait linear_sumcheck::data_structures::protocol::Protocol

pub trait Protocol: Sized {
    type InboundMessage: Message;
    type OutBoundMessage: Message;
    type Error: Error + From<Error> + Display;
    fn current_round(&self) -> Result<u32, Self::Error>;

    fn is_active(&self) -> bool;

    fn get_message(
        &self, 
        round: u32
    ) -> Result<Self::OutBoundMessage, Self::Error>;

    fn push_message(
        &mut self, 
        msg: &Self::InboundMessage
    ) -> Result<(), Self::Error>;

    fn get_latest_message(
        &mut self
    ) -> Result<Self::OutBoundMessage, Self::Error> { ... }
}

Interactive protocol

Associated Types

type InboundMessage: Message

Message sent to this protocol from others.

type OutBoundMessage: Message

Message sent from this protocol to others.

type Error: Error + From<Error> + Display

Type of error

Loading content...
Required methods

fn current_round(&self) -> Result<u32, Self::Error>

Get current round. If the protocol is not active (e.g. in accepted or rejected status), current_round should return an InvalidOperationError.

fn is_active(&self) -> bool

If this method is true, user can push message to this protocol at this time.

Note: inactive doesn't mean user cannot get latest message.

fn get_message(&self, round: u32) -> Result<Self::OutBoundMessage, Self::Error>

Get message from this protocol at round

If round < current round, the protocol simply returns message from cache. If the protocol does not support cache, this protocol will return an error.

fn push_message(
&mut self,
msg: &Self::InboundMessage
) -> Result<(), Self::Error>

Push message to this protocol

Protocol goes to next round, or become inactive.

Loading content...
Provided methods

fn get_latest_message(&mut self) -> Result<Self::OutBoundMessage, Self::Error>

get message from this protocol at current round

Repeatedly calling get_latest_message without pushing lead to same result.

Loading content...