Trait linear_sumcheck::gkr_round_sumcheck::GKRRoundClaim

pub trait GKRRoundClaim<F: Field>: Clone + CanonicalSerialize + CanonicalDeserialize {
    fn g(&self) -> Vec<F>;

    fn dim(&self) -> u32;

    fn asserted_sum(&self) -> F;
}

fixed point g, and the sum of the GKR round function.

Required methods

fn g(&self) -> Vec<F>

a vector representing fixed point g on the hypercube

fn dim(&self) -> u32

the dimension of the GKR function to be proved (number variables of g)

fn asserted_sum(&self) -> F

asserted sum of the GKR round function when g is fixed

Loading content...