Trait linear_sumcheck::ml_sumcheck::MLSumcheck

pub trait MLSumcheck<F> where
    F: Field, {
    type Claim: MLSumcheckClaim<F>;
    type Proof: Clone + CanonicalSerialize + CanonicalDeserialize;
    type Error: Error + From<Error>;
    type SubClaim: MLSumcheckSubclaim<F>;
    fn generate_claim_and_proof<P: MLExtension<F>>(
        poly: &[P]
    ) -> Result<(Self::Claim, Self::Proof), Self::Error>;

    fn verify_proof(
        claim: &Self::Claim, 
        proof: &Self::Proof
    ) -> Result<Self::SubClaim, Self::Error>;
}

ML sumcheck protocol

Associated Types

type Claim: MLSumcheckClaim<F>

Claim of the sum

type Proof: Clone + CanonicalSerialize + CanonicalDeserialize

proof of the sum

type Error: Error + From<Error>

error type

type SubClaim: MLSumcheckSubclaim<F>

subclaim outputted: if the claim is true, the subclaim should be true

Loading content...
Required methods

fn generate_claim_and_proof<P: MLExtension<F>>(
poly: &[P]
) -> Result<(Self::Claim, Self::Proof), Self::Error>

Calculate the sum of the polynomial and generate the proof.

poly: product of multilinear functions represented by an iterator of multilinear function

fn verify_proof(
claim: &Self::Claim,
proof: &Self::Proof
) -> Result<Self::SubClaim, Self::Error>

verify if proof correctly proves the claim

Loading content...