Trait linear_sumcheck::ml_sumcheck::MLSumcheckSubclaim

pub trait MLSumcheckSubclaim<F>: Clone + CanonicalSerialize + CanonicalDeserialize {
    fn evaluation_point(&self) -> Vec<F>;

    fn expected_evaluations(&self) -> F;
}

subclaim of the MLSumcheck

The subclaim consists of a point on the dimension-n hypercube and the expected evaluation of that point in the finite field.

Required methods

fn evaluation_point(&self) -> Vec<F>

a point on dimension-n hypercube, where n is the number of variables of the polynomial

fn expected_evaluations(&self) -> F

get expected evaluation on the point in the finite field

Loading content...